requirements.txt
django==4.1.3
celery==5.2.7
redis==4.4.0
sqlalchemy==1.4.44
djangorestframework==3.14.0
# ffmpeg-python==0.2.0
ffmpeg==1.4
